+++ src/AgendaGlow/App.js
@@ -1,6 +1,6 @@
 import React from 'react';
-import { AuthProvider } from './src/src/context/AuthContext';
-import AppRoutes from './src/src/navigations/AppRoutes';
+import { AuthProvider } from './src/context/AuthContext';
+import AppRoutes from './src/navigations/AppRoutes';
 
 export default function App() {
   return (
