+++ src/AgendaGlow/src/screens/FuncionarioEditar.js
@@ -1,29 +1,29 @@
-// filepath: c:\Users\Melissa\Desktop\psg-ads-n-tiam-2025-2-tiam_grupo_1_salaodebeleza\src\AgendaGlow\src\screens\FuncionarioEditar.js
 import React, { useState, useContext } from 'react';
-import { View, StyleSheet, ScrollView, Text, Modal } from 'react-native'; 
+import { View, StyleSheet, ScrollView, Text, Modal, TouchableOpacity } from 'react-native';
 import Header from '../components/Header';
 import Input from '../components/Input';
 import Button from '../components/Button';
 import { theme } from '../styles/theme';
-import { 
-  deleteFuncionario, 
-  updateFuncionarioPassword, 
-  updateFuncionarioComAuth 
+import { Ionicons } from '@expo/vector-icons';
+import {
+  deleteFuncionario,
+  updateFuncionarioPassword,
+  updateFuncionarioComAuth,
 } from '../services/funcionarioService';
 import { auth } from '../database/firebase';
 import { AuthContext } from '../context/AuthContext';
 
 export default function FuncionarioEditar({ navigation, route }) {
-
   const { user: userLogado } = useContext(AuthContext);
-
-  const funcionario = route && route.params && route.params.funcionario ? route.params.funcionario : null;
+  const funcionario = route?.params?.funcionario || null;
 
   if (!funcionario) {
     return (
       <View style={[styles.container, { justifyContent: 'center', alignItems: 'center' }]}>
-        <Header userName="Usuário" />
-        <Text style={{ color: theme.colors.text, fontSize: 16 }}>Nenhum funcionário selecionado para edição.</Text>
+        <Header pageTitle={"EDITAR PROFISSIONAL"} />
+        <Text style={{ color: theme.colors.text, fontSize: 16 }}>
+          Nenhum profissional selecionado para edição.
+        </Text>
         <Button title="Voltar" onPress={() => navigation.goBack()} style={{ marginTop: 16 }} />
       </View>
     );
@@ -45,21 +45,39 @@ export default function FuncionarioEditar({ navigation, route }) {
   const [errorSave, setErrorSave] = useState(null);
   const [errorPassword, setErrorPassword] = useState(null);
   const [errorDelete, setErrorDelete] = useState(null);
+  const [errorSenhaModal, setErrorSenhaModal] = useState(null);
 
   const [modalVisible, setModalVisible] = useState(false);
   const [modalPasswordVisible, setModalPasswordVisible] = useState(false);
+  const [modalConfirmVisible, setModalConfirmVisible] = useState(false);
+
+  // Modal de feedback padrão
+  const [modalFeedbackVisible, setModalFeedbackVisible] = useState(false);
+  const [modalFeedbackTitle, setModalFeedbackTitle] = useState('');
+  const [modalFeedbackMessage, setModalFeedbackMessage] = useState('');
+  const [modalFeedbackType, setModalFeedbackType] = useState('success'); 
 
   const providerId = userLogado?.providerData[0]?.providerId || 'password';
   const isPasswordUser = providerId === 'password';
   const isEditingSelf = userLogado && userLogado.uid === funcionario.uid;
 
-  // Função chamada ao tocar em "Salvar Dados" (abre modal se precisar confirmar senha)
+  const showFeedbackModal = (title, message, type = 'success') => {
+    setModalFeedbackTitle(title);
+    setModalFeedbackMessage(message);
+    setModalFeedbackType(type);
+    setModalFeedbackVisible(true);
+  };
+
+  const closeFeedbackModal = () => {
+    setModalFeedbackVisible(false);
+    if (modalFeedbackType === 'success') navigation.goBack();
+  };
+
+  // --- Salvar Dados ---
   const handleSalvarPress = () => {
     setErrorSave(null);
-
-    // Validação de campos obrigatórios e formato de e-mail
-    if (!nome || !email) {
-      setErrorSave('Nome e E-mail são obrigatórios.');
+    if (!nome || !email || !telefone ) {
+      setErrorSave('Preencha os campos obrigatórios.');
       return;
     }
     if (!email.includes('@') || !email.includes('.')) {
@@ -67,51 +85,39 @@ export default function FuncionarioEditar({ navigation, route }) {
       return;
     }
 
-    if (isPasswordUser) {
-      // Abre modal para solicitar senha de confirmação
-      setModalVisible(true);
-      return;
-    }
-
-    // Se não precisa de senha para confirmar, chama a função de edição diretamente
-    handleEditar();
+    if (isPasswordUser) setModalVisible(true);
+    else handleEditar();
   };
 
-  // --- Função Salvar Dados (executa atualização, espera senhaConfirmacao se aplicável) ---
   const handleEditar = async () => {
-    setErrorSave(null);
+    setErrorSenhaModal(null);
 
-    // Se for usuário com provider "password", exige senhaConfirmacao preenchida
     if (isPasswordUser && !senhaConfirmacao) {
-      setErrorSave('Digite sua senha atual para confirmar as alterações.');
+      setErrorSenhaModal('Digite sua senha atual para confirmar.');
       return;
     }
 
     setLoadingSave(true);
     const dadosAtualizados = { nome, telefone, email: email.toLowerCase() };
-
     const result = await updateFuncionarioComAuth(
-      funcionario.id, 
-      dadosAtualizados, 
+      funcionario.id,
+      dadosAtualizados,
       isPasswordUser ? senhaConfirmacao : null
     );
-
     setLoadingSave(false);
-    setSenhaConfirmacao(''); 
-    setModalVisible(false);
+    setSenhaConfirmacao('');
 
     if (result.success) {
-      alert('Sucesso: Dados atualizados com sucesso!');
-      navigation.goBack(); 
-    } else {
-      setErrorSave(result.message || 'Falha ao atualizar dados.');
-    }
-  };
+    setModalVisible(false);
+    showFeedbackModal('Sucesso', 'Dados atualizados com sucesso!', 'success');
+  } else {
+    showFeedbackModal('Erro', result.message || 'Falha ao atualizar dados.', 'error');
+  }
+};
 
-  // --- Função Alterar Senha ---
+  // --- Alterar Senha ---
   const handleAlterarSenha = async () => {
-    setErrorPassword(null); 
-    
+    setErrorPassword(null);
     if (!senhaAtual || !novaSenha || !confirmarSenha) {
       setErrorPassword('Preencha todos os campos de senha.');
       return;
@@ -130,67 +136,67 @@ export default function FuncionarioEditar({ navigation, route }) {
     setLoadingPassword(false);
 
     if (result.success) {
-      alert('Sucesso: Senha alterada com sucesso!');
+      showFeedbackModal('Sucesso', 'Senha alterada com sucesso!', 'success');
       setSenhaAtual('');
       setNovaSenha('');
       setConfirmarSenha('');
       setModalPasswordVisible(false);
     } else {
-      setErrorPassword(result.message || 'Ocorreu uma falha.');
+      showFeedbackModal('Erro', result.message || 'Falha ao alterar senha.', 'error');
     }
   };
 
-  // --- Função Excluir Conta (A ser corrigida no servidor no futuro) ---
-  const handleExcluir = async () => {
-    setErrorDelete(null); 
-    
-    const deveExcluir = window.confirm(
-      `Confirmar Exclusão: Tem certeza que deseja excluir ${funcionario.nome}? Esta ação é irreversível.`
-    );
-    
-    if (deveExcluir) {
-      setLoadingDelete(true);
-      const result = await deleteFuncionario(funcionario.uid, funcionario.id); 
-      setLoadingDelete(false);
-      
-      if (result.success) {
-        alert('Sucesso: Funcionário excluído com sucesso.');
-        if (isEditingSelf) { auth.signOut(); } 
-        else { navigation.goBack(); }
-      } else {
-        setErrorDelete(result.message || 'Falha ao excluir funcionário.');
-      }
+  // --- Excluir Conta (soft delete) ---
+  const handleExcluirConfirmado = async () => {
+    setModalConfirmVisible(false);
+    setErrorDelete(null);
+    setLoadingDelete(true);
+
+    const result = await deleteFuncionario(funcionario.uid, funcionario.id);
+    setLoadingDelete(false);
+
+    if (result.success) {
+      showFeedbackModal('Sucesso', 'Conta excluída com sucesso.', 'success');
+      if (isEditingSelf) auth.signOut();
+    } else {
+      showFeedbackModal('Erro', result.message || 'Falha ao excluir conta.', 'error');
     }
   };
 
   return (
     <View style={styles.container}>
-      <Header userName={nome.split(' ')[0]} /> 
+      <Header pageTitle={"EDITAR PROFISSIONAL"}/>
       <ScrollView contentContainerStyle={styles.content}>
         <Text style={styles.title}>Alterar meus dados</Text>
-  
-        <Input placeholder="Nome" value={nome} onChangeText={setNome} />
-        <Input placeholder="Telefone" value={telefone} onChangeText={setTelefone} keyboardType="phone-pad" />
-        <Input placeholder="E-mail" value={email} onChangeText={setEmail} keyboardType="email-address" autoCapitalize="none" />
 
-        {/* REMOVIDO: campo inline de confirmação de senha.
-            Agora a senha é solicitada dentro de um modal ao clicar em "Salvar Dados" */}
-        
-        {errorSave && (
-          <Text style={styles.errorMessage}>{errorSave}</Text>
-        )}
-        
+        <Input placeholder="Nome" value={nome} onChangeText={setNome} />
+        <Input
+          placeholder="Telefone"
+          value={telefone}
+          onChangeText={setTelefone}
+          keyboardType="phone-pad"
+          maxLength={17}
+        />
+        <Input
+          placeholder="E-mail"
+          value={email}
+          onChangeText={setEmail}
+          keyboardType="email-address"
+          autoCapitalize="none"
+        />
+
+        {errorSave && <Text style={styles.errorMessage}>{errorSave}</Text>}
+
         <Button
           title={loadingSave ? 'Salvando...' : 'Salvar Dados'}
           onPress={handleSalvarPress}
           style={styles.saveButton}
           disabled={loadingSave || loadingDelete || loadingPassword}
         />
-        
+
         {isEditingSelf && isPasswordUser && (
           <>
             <Text style={styles.sectionTitle}>Outras Opções</Text>
-
             <Button
               title="Alterar Senha"
               onPress={() => {
@@ -202,39 +208,70 @@ export default function FuncionarioEditar({ navigation, route }) {
             />
           </>
         )}
-        
+
         {isEditingSelf && (
-           <View style={{ marginTop: theme.spacing.large, width: '100%' }}>
-              {errorDelete && (
-                <Text style={styles.errorMessage}>{errorDelete}</Text>
-              )}
-              
-              <Button
-                title={loadingDelete ? 'Excluindo...' : 'Excluir Conta'}
-                onPress={handleExcluir}
-                style={styles.primary} 
-                disabled={loadingSave || loadingDelete || loadingPassword}
-              />
-              <Text style={styles.deleteWarning}>Atenção: Excluir a conta é uma ação permanente.</Text>
-              
+          <View style={{ width: '100%' }}>
+            {errorDelete && <Text style={styles.errorMessage}>{errorDelete}</Text>}
+
+            <Button
+              title={loadingDelete ? 'Excluindo...' : 'Excluir Conta'}
+              onPress={() => setModalConfirmVisible(true)}
+              style={styles.saveButton}
+              disabled={loadingSave || loadingDelete || loadingPassword}
+            />
+            <Text style={styles.deleteWarning}>
+              Atenção: Excluir a conta é uma ação permanente.
+            </Text>
           </View>
         )}
-
       </ScrollView>
 
-      {/* Modal para solicitar senha de confirmação ao salvar (apenas para usuários com provider password) */}
+      {/* --- Modal de confirmação de exclusão --- */}
       <Modal
-        visible={modalVisible}
+        visible={modalConfirmVisible}
         transparent
-        animationType="slide"
-        onRequestClose={() => setModalVisible(false)}
+        animationType="fade"
+        onRequestClose={() => setModalConfirmVisible(false)}
       >
-        <View style={styles.modalContainer}>
-          <View style={styles.modalContent}>
-            <Text style={styles.modalTitle}>Confirme sua senha</Text>
-            <Text style={{ color: theme.colors.text, fontSize: 13, marginBottom: 8 }}>
-              Digite sua senha atual para confirmar as alterações.
+        <View style={styles.modalOverlay}>
+          <View style={styles.modalContainer}>
+            <View style={styles.modalHeader}>
+              <Text style={styles.modalTitle}>Confirmar Exclusão</Text>
+              <TouchableOpacity onPress={() => setModalConfirmVisible(false)}>
+                <Ionicons name="close" size={24} color={theme.colors.text} />
+              </TouchableOpacity>
+            </View>
+            <Text style={styles.modalMessage}>
+              Tem certeza de que deseja excluir sua conta? Esta ação não pode ser desfeita.
             </Text>
+
+            <View style={{ flexDirection: 'row', justifyContent: 'flex-end', gap: 10 }}>
+              <Button
+                title="Cancelar"
+                onPress={() => setModalConfirmVisible(false)}
+                style={{ backgroundColor: theme.colors.cancel }}
+              />
+              <Button
+                title={loadingDelete ? 'Excluindo...' : 'Confirmar'}
+                onPress={handleExcluirConfirmado}
+                disabled={loadingDelete}
+              />
+            </View>
+          </View>
+        </View>
+      </Modal>
+
+      {/* Modal de senha para salvar */}
+      <Modal visible={modalVisible} transparent animationType="fade" onRequestClose={() => setModalVisible(false)}>
+        <View style={styles.modalOverlay}>
+          <View style={styles.modalContainer}>
+            <View style={styles.modalHeader}>
+              <Text style={styles.modalTitle}>Confirme sua senha</Text>
+              <TouchableOpacity onPress={() => setModalVisible(false)}>
+                <Ionicons name="close" size={24} color={theme.colors.text} />
+              </TouchableOpacity>
+            </View>
+            <Text style={styles.modalMessage}>Digite sua senha atual para confirmar.</Text>
             <Input
               placeholder="Senha Atual"
               value={senhaConfirmacao}
@@ -242,16 +279,12 @@ export default function FuncionarioEditar({ navigation, route }) {
               secureTextEntry
             />
 
+            {/* Exibe mensagem de erro dentro do modal */}
+            {errorSenhaModal && (
+              <Text style={styles.errorMessage}>{errorSenhaModal}</Text>
+            )}
+
             <View style={styles.modalButtons}>
-              <Button
-                title="Cancelar"
-                onPress={() => {
-                  setSenhaConfirmacao('');
-                  setModalVisible(false);
-                }}
-                style={{ marginRight: 8, backgroundColor: theme.colors.container2 }}
-                disabled={loadingSave}
-              />
               <Button
                 title={loadingSave ? 'Salvando...' : 'Confirmar'}
                 onPress={handleEditar}
@@ -262,41 +295,21 @@ export default function FuncionarioEditar({ navigation, route }) {
         </View>
       </Modal>
 
-      {/* Modal para alteração de senha */}
-      <Modal
-        visible={modalPasswordVisible}
-        transparent
-        animationType="slide"
-        onRequestClose={() => setModalPasswordVisible(false)}
-      >
-        <View style={styles.modalContainer}>
-          <View style={styles.modalContent}>
-            <Text style={styles.modalTitle}>Alterar Senha</Text>
-            <Text style={{ color: theme.colors.text, fontSize: 13, marginBottom: 8 }}>
-              Informe sua senha atual e a nova senha.
-            </Text>
-
+      {/* Modal de senha para alteração */}
+      <Modal visible={modalPasswordVisible} transparent animationType="fade" onRequestClose={() => setModalPasswordVisible(false)}>
+        <View style={styles.modalOverlay}>
+          <View style={styles.modalContainer}>
+            <View style={styles.modalHeader}>
+              <Text style={styles.modalTitle}>Alterar Senha</Text>
+              <TouchableOpacity onPress={() => setModalPasswordVisible(false)}>
+                <Ionicons name="close" size={24} color={theme.colors.text} />
+              </TouchableOpacity>
+            </View>
             <Input placeholder="Senha Atual" value={senhaAtual} onChangeText={setSenhaAtual} secureTextEntry />
             <Input placeholder="Nova Senha (mín. 6 caracteres)" value={novaSenha} onChangeText={setNovaSenha} secureTextEntry />
             <Input placeholder="Confirmar Nova Senha" value={confirmarSenha} onChangeText={setConfirmarSenha} secureTextEntry />
-
-            {errorPassword && (
-              <Text style={styles.errorMessage}>{errorPassword}</Text>
-            )}
-
+            {errorPassword && <Text style={styles.errorMessage}>{errorPassword}</Text>}
             <View style={styles.modalButtons}>
-              <Button
-                title="Cancelar"
-                onPress={() => {
-                  setSenhaAtual('');
-                  setNovaSenha('');
-                  setConfirmarSenha('');
-                  setErrorPassword(null);
-                  setModalPasswordVisible(false);
-                }}
-                style={{ marginRight: 8, backgroundColor: theme.colors.container2 }}
-                disabled={loadingPassword}
-              />
               <Button
                 title={loadingPassword ? 'Alterando...' : 'Confirmar'}
                 onPress={handleAlterarSenha}
@@ -306,64 +319,42 @@ export default function FuncionarioEditar({ navigation, route }) {
           </View>
         </View>
       </Modal>
+
+      {/* Modal de feedback padrão */}
+      <Modal visible={modalFeedbackVisible} transparent animationType="fade" onRequestClose={closeFeedbackModal}>
+        <View style={styles.modalOverlay}>
+          <View style={styles.modalContainer}>
+            <View style={styles.modalHeader}>
+              <Text style={styles.modalTitle}>{modalFeedbackTitle}</Text>
+              <TouchableOpacity onPress={closeFeedbackModal}>
+                <Ionicons name="close" size={24} color={theme.colors.text} />
+              </TouchableOpacity>
+            </View>
+            <Text style={styles.modalMessage}>{modalFeedbackMessage}</Text>
+            <View style={styles.modalButtons}>
+              <Button title="OK" onPress={closeFeedbackModal} />
+            </View>
+          </View>
+        </View>
+      </Modal>
     </View>
   );
 }
 
+// --- ESTILOS ---
 const styles = StyleSheet.create({
   container: { flex: 1, backgroundColor: theme.colors.background },
   content: { padding: theme.spacing.large },
   title: { fontSize: 20, fontWeight: '700', color: theme.colors.text, marginBottom: theme.spacing.medium },
-  sectionTitle: { fontSize: 16, fontWeight: '700', color: theme.colors.primary, marginTop: theme.spacing.large, marginBottom: theme.spacing.small, alignSelf: 'flex-start' },
-  label: { fontSize: 14, color: theme.colors.text, fontWeight: '500', marginLeft: 4, marginTop: 8 },
-  labelConfirm: { fontSize: 14, color: theme.colors.text, fontWeight: '700', marginLeft: 4, marginTop: 12 },
-  
-  errorMessage: {
-    color: '#D9534F',
-    fontSize: 14,
-    fontWeight: '600',
-    textAlign: 'center',
-    marginTop: 8,
-    marginBottom: 4,
-  },
-  inputDisabled: { 
-    backgroundColor: theme.colors.container3 || '#FCEDED', 
-    color: theme.colors.textInput 
-  },
-  saveButton: { marginTop: theme.spacing.medium },
-  deleteButton: { 
-    marginTop: theme.spacing.small, 
-    backgroundColor: theme.colors.error || '#D9534F' 
-  },
-  deleteWarning: { 
-    color: theme.colors.error || '#D9534F', 
-    textAlign: 'center', 
-    marginBottom: 8, 
-    fontSize: 12 
-  },
+  sectionTitle: { fontSize: 16, fontWeight: '700', color: theme.colors.primary, marginTop: theme.spacing.large, marginBottom: theme.spacing.small },
+  errorMessage: { color: theme.colors.primary, fontSize: 14, fontWeight: '600', textAlign: 'center', marginTop: 8 },
+  deleteWarning: { color: theme.colors.textSecondary || '#777', textAlign: 'center', fontSize: 12, marginTop: 6 },
 
-  /* Styles do modal */
-  modalContainer: {
-    flex: 1,
-    backgroundColor: 'rgba(0,0,0,0.5)',
-    justifyContent: 'center',
-    paddingHorizontal: 24,
-  },
-  modalContent: {
-    backgroundColor: theme.colors.background,
-    borderRadius: 10,
-    padding: theme.spacing.large,
-    elevation: 5,
-  },
-  modalTitle: {
-    fontSize: 18,
-    fontWeight: '700',
-    color: theme.colors.text,
-    marginBottom: theme.spacing.small,
-  },
-  modalButtons: {
-    flexDirection: 'row',
-    justifyContent: 'flex-end',
-    marginTop: theme.spacing.medium,
-  },
+  // Modal padrão
+  modalOverlay: { flex: 1, backgroundColor: 'rgba(0,0,0,0.5)', justifyContent: 'center', alignItems: 'center', paddingHorizontal: 20 },
+  modalContainer: { backgroundColor: theme.colors.white, borderRadius: 16, width: '100%', maxWidth: 400, padding: 20, elevation: 10 },
+  modalHeader: { flexDirection: 'row', justifyContent: 'space-between', alignItems: 'center', marginBottom: 12 },
+  modalTitle: { fontSize: 18, fontWeight: '700', color: theme.colors.text },
+  modalMessage: { fontSize: 16, color: theme.colors.text, marginBottom: 20 },
+  modalButtons: { alignItems: 'flex-end' },
 });
